Decompile func0f161c08

+++ src/include/types.h
@@ -4679,7 +4679,7 @@ struct portalthing {
 struct var800a4ccc { // related to portals
 	struct coord coord;
 	f32 unk0c;
-	u32 unk10;
+	f32 unk10;
 };
 
 struct trainingdata {
